src/console/requirements.txt
django<6
jupyter
opentelemetry-sdk
opentelemetry-exporter-otlp